+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Fixed: Option to specify the Pester version for the Invoke-BuildIsolated command to prevent Pester module version mismatch
+
 ## 3.2.2 - 2023-01-07
 
 * Fixed: Remove assembly files from schema check